Use command pointer checks in SE setters

+++ src/RedSound/RedCommand.cpp
@@ -564,7 +564,7 @@ void SetSeVolume(int seId, int volume, int frameCount, int mode)
 	track = ((RedSoundCONTROL*)p_SoundControlBuffer)[REDSOUND_CONTROL_SE].m_tracks;
 
 	do {
-		if (((u32)track->m_command != 0) && ((seId < 0) || (track->m_seId == seId))) {
+		if ((track->m_command != 0) && ((seId < 0) || (track->m_seId == seId))) {
 			int delta = volume - track->m_mixVolume;
 			delta /= frameCount;
 			track->m_mixVolumeAdd = delta;
@@ -599,7 +599,7 @@ void SetSePan(int seId, int pan, int frameCount)
 	track = ((RedSoundCONTROL*)p_SoundControlBuffer)[REDSOUND_CONTROL_SE].m_tracks;
 
 	do {
-		if (((u32)track->m_command != 0) && ((seId < 0) || (track->m_seId == seId))) {
+		if ((track->m_command != 0) && ((seId < 0) || (track->m_seId == seId))) {
 			int delta = pan - track->m_pan;
 			delta /= frameCount;
 			track->m_panAdd = delta;
@@ -633,7 +633,7 @@ void SetSePitch(int seId, int pitch, int frameCount)
 	track = ((RedSoundCONTROL*)p_SoundControlBuffer)[REDSOUND_CONTROL_SE].m_tracks;
 
 	do {
-		if (((u32)track->m_command != 0) && ((seId < 0) || (track->m_seId == seId))) {
+		if ((track->m_command != 0) && ((seId < 0) || (track->m_seId == seId))) {
 			int delta = pitch - track->m_pitch;
 			track->m_pitchAdd = delta / frameCount;
 			track->m_pitchDelta = frameCount;
